VitalHealth/healthbot_v3/requirements.txt
# ── Web framework ─────────────────────────────────────────────────────────────
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.12

# ── Data validation ───────────────────────────────────────────────────────────
pydantic==2.9.2

# ── Embeddings ────────────────────────────────────────────────────────────────
sentence-transformers==3.1.1
numpy==1.26.4

# ── LLM (GGUF runtime) ───────────────────────────────────────────────────────
# If you have a GPU with CUDA:
#   pip install llama-cpp-python --extra-index-url https://abetlen.github.io/llama-cpp-python/whl/cu121
# CPU only (default):
llama-cpp-python==0.3.1

# ── Document parsing ──────────────────────────────────────────────────────────
pdfplumber==0.11.4
Pillow==10.4.0

# ── OCR (PaddleOCR is optional but strongly recommended for prescriptions) ────
# Uncomment if you want PaddleOCR support (primary OCR engine):
# paddlepaddle==2.6.1
# paddleocr==2.8.1

# pytesseract is the fallback OCR engine (always install this)
pytesseract==0.3.13
